Hard Coded
GUI_Main is the main class to be ran. 
Display_Page:  holds the the menu bar for the overall page. Bellow it is where the scenes will change. 
Log in page will still open first and depending on what is entered a scene will open. 

Host_2 is a border pane. Each section is broken up into a different class then displayed. 
Have started on the event page, so we can have one page to do custom events. 

Css.css does not really have a use yet.