Hard coded
GUI_Main is the main runner class .
Display_Page holds the background and the menu bar that will get changed. 
the center of the Display_Page will change scenes 

Login is the first to be displayed. Depending on who signs in, the Display_page will show either Host, Manager, Or Kitchen